// components/selectors/types.ts
export type periodType = 'rok' | 'półrocze' | 'kwartał' | 'miesiąc' | 'konkretne';

export type halfYearType = 'I półrocze' | 'II półrocze';

export type quarterTypes = 'I_kwrtał' | 'II_kwrtał' | 'III_kwrtał' | 'IV_kwrtał';

export type monthTypes =
  | 'styczeń'
  | 'luty'
  | 'marzec'
  | 'kwiecień'
  | 'maj'
  | 'czerwiec'
  | 'lipiec'
  | 'sierpień'
  | 'wrzesień'
  | 'październik'
  | 'listopad'
  | 'grudzień';